T#807 P3-C: Extract guest module from server.ts — FINAL P3 sub-task
Sub-task C of T#786 Phase 3. Final extraction.

Extracts 13 guest-facing handlers (~598 LoC) + 1 helper function
(getGuestDisplayName) + 1 rate-limit Map state into src/guest/routes.ts.

## Changes
- New: src/guest/routes.ts (590L) — registerGuestRoutes(app, sqlite, helpers)
  with GuestHelpers DI (wsBroadcast, withRetry, getTmuxStatus,
  normalizeAvatarUrl, uploadsDir)
- server.ts: 2505L → 1908L (-597L)

## Handlers
- GET /api/guest/dashboard, /api/guest/threads, /api/guest/thread/:id, /api/guest/pack, /api/guest/profile, /api/guest/dm/:from/:to
- POST /api/guest/thread/:id/message, /api/guest/thread, /api/guest/dm, /api/guest/change-password, /api/guest/reset-password, /api/guest/avatar
- PATCH /api/guest/profile

Module-local state: passwordChangeAttempts rate-limit Map (5 attempts /
15 min per guest username) preserved verbatim.

## Cumulative for T#786 P3 (all 10 sub-tasks)

server.ts: 4428L → 1908L (-2520L, **57%% stripped**)

Modules created (10):
- src/settings/routes.ts (T#798)
- src/remote/routes.ts (T#799)
- src/queue/routes.ts (T#800)
- src/supersede/routes.ts (T#801)
- src/audit/routes.ts (T#802)
- src/teams/routes.ts (T#803)
- src/files/routes.ts (T#804)
- src/inbox/routes.ts (T#805)
- src/knowledge/routes.ts (T#806)
- src/guest/routes.ts (T#807) — this PR

Target was ~500L; 1908L is the final landing. Remaining content is
imports, middleware, DB init, module wiring, WebSocket, SPA fallback,
graceful shutdown — proper server.ts entry-point shape now.

## Test plan
- [x] bun build clean (Norm #84 Layer 1)
- [x] Module registration before SPA fallback
- [ ] Post-deploy smoke

Reviewer: @bertus per task assignment (security-touching — guest auth, rate
limits, injection scan).
